tasks/load-opa-assessments/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.11.4
